Skeleton Structure
We will be using poetry package dependency manager to handle our external libraries such as Vidgear and OpenCV.

This commit serves as the purpose of creating a structure for the home controller system such that other members of the team can work on a separate branch and use the same structure.

I have installed the foollowing packages:
* Black - formatting library to allign our python code to the industry standards.
* poetrify - used for compilling the dependencies into a requirements.txt that will be used to install our packages on deploy.

I have included a ReadMe that explains how to use poetry and integrate it with respective IDEs for development.