Movement Detection
There were several bugs in the previous implementation such that adding client were not logically correct. This was preventing any form of video stream to be produced to the terminal when running the server and client. Hence, by resolving these issues, I was able to view the live stream with no problem.

Since the logical errors were cleared for the live streaming, it allowed me to focus on the logical expressions in the movement detection. The following is done to achieve movement:
* Moving average for a specifc client is saved in the Client class in server.py. This is used each time the frame is sent from the client.py (IP cameras). A threshold is made (ceil) that defines when movement is acceptable to trigger the intruder detection module.