Movement Detection optional
The movement detection need not be tested when testing for sending a frame. Hence, this detect_movement parameter in run will allow for testing only the recieving of a frame. Hence, now the movment detection is not covered by the previous test of checking if the server receieves the frame.

TODO: with the updated parameter, now please test whether there is movement by sendind vid frames.