Squashed commit of the following
commit c687475d 
Author: UshirRaval <UshirRaval> 
Date: Thu Jun 11 2020 16:57:23 GMT+0200 (SAST) 

    Conflicts resolved in the console.

commit 9bb11efe 
Author: u17080534 <55683117+u17080534@users.noreply.github.com> 
Date: Thu Jun 11 2020 16:06:46 GMT+0200 (SAST) 

    Merge branch 'IP_Camera_Configuration' of https://git-codecommit.eu-west-1.amazonaws.com/v1/repos/Home_Control_Panel into IP_Camera_Configuration


commit 802d41c6 
Author: u17080534 <55683117+u17080534@users.noreply.github.com> 
Date: Thu Jun 11 2020 16:06:29 GMT+0200 (SAST) 

    Update test_server.py


commit 1cf06569 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 16:06:20 GMT+0200 (SAST) 

    Merge branch 'IP_Camera_Configuration' of https://git-codecommit.eu-west-1.amazonaws.com/v1/repos/Home_Control_Panel into IP_Camera_Configuration


commit d4151dac 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 16:00:57 GMT+0200 (SAST) 

    Test Pass


commit c5698dd8 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 16:05:31 GMT+0200 (SAST) 

    Toggle movement detection to not be triggered with this test

This test only detects if the frame is received by the server, and should not detect movement.


commit 32648f3f 
Author: u17080534 <55683117+u17080534@users.noreply.github.com> 
Date: Thu Jun 11 2020 15:49:14 GMT+0200 (SAST) 

    Added Server Tests


commit 4a17a703 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 16:00:57 GMT+0200 (SAST) 

    Test Pass

The previous tests were failing due to the detect_movement parameter not being accounted for in the server.run function. Now it toggles the movement detection.

commit ff28e28e 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 14:59:49 GMT+0200 (SAST) 

    Movement Detection optional

The movement detection need not be tested when testing for sending a frame. Hence, this detect_movement parameter in run will allow for testing only the recieving of a frame. Hence, now the movment detection is not covered by the previous test of checking if the server receieves the frame.

TODO: with the updated parameter, now please test whether there is movement by sendind vid frames.


commit 122b57c3 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 14:35:54 GMT+0200 (SAST) 

    Vidgear Corner Case

for some reason, if the code added in server.py is not there, then the montage does not display the live video.


commit 7bc034e2 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 14:16:05 GMT+0200 (SAST) 

    Black formatting

formattting changes for python industry standard code format


commit 4250346a 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 14:07:27 GMT+0200 (SAST) 

    Test Server

Server has 78% test coverage. Run the tests in coverage mode to see what still needs to be tested.


commit b30ea7ad 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jun 11 2020 03:44:50 GMT+0200 (SAST) 

    Movement Detection

There were several bugs in the previous implementation such that adding client were not logically correct. This was preventing any form of video stream to be produced to the terminal when running the server and client. Hence, by resolving these issues, I was able to view the live stream with no problem.

Since the logical errors were cleared for the live streaming, it allowed me to focus on the logical expressions in the movement detection. The following is done to achieve movement:
* Moving average for a specifc client is saved in the Client class in server.py. This is used each time the frame is sent from the client.py (IP cameras). A threshold is made (ceil) that defines when movement is acceptable to trigger the intruder detection module.


commit ea2005cf 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Wed Jun 10 2020 08:01:35 GMT+0200 (SAST) 

    Created Multi-server for hosting mutliple ip camera clients

* The server class is the main controller that handles adding ip cameras to the system and handles movement detection.

* I currently have the multi-threaded implementation done using Netgear as a Vidgear gear. This, coupled with OpenCV allows for the server to handle a multi-threaded environment.

** The server stores local Clients in a dictionary where each client has a unique port that connects to the server. This is assigned using the add function.

TODO:
** I have not managed to get the movement detection video to display to the screen when an movement is analysed in the Client class in server.py. I have suspition that it because I used a reference for caclulating moement detection from an outdated source. So I had to find the relavant functions that are now discontinued :arrow_lower_right:

However, the tcp server is set up in a modular where each sub function is extrapolated into functions that needs to be tested for.

** Unit teting/Pytest: Test the individual functions in the Server class. Since the server uses live streams, you may need to provide your own client snippet that sends a precompiled video and tests your functions by sending the video on an appropriate port to the server - This is used for testing the endpoints such as:
*** sending the movement detection frames to the intruder detection lamdba function (to come)
*** SQS queue for storing movement and reandom video streams sent from the local server.

Acceptance Criteria acheived:
* Create multi-threaded server to handle multiple ip cameras for a specific owner.
* 80 % for movement detection. please visit the following links for understanding my approach:
** Multi-threaded server: https://github.com/abhiTronix/vidgear/wiki/Multi-Server-Mode-for-NetGear-API
** Movement detection:
https://www.pyimagesearch.com/2015/05/25/basic-motion-detection-and-tracking-with-python-and-opencv/ ;
http://www.robindavid.fr/opencv-tutorial/motion-detection-with-opencv.html ;
https://github.com/mattwilliamson/Motion-Tracker

NOTE:
to view live streams, use the openCV command cv.immshow("Video Title", frame) to open it up on your terminal


commit 3c1f6417 
Author: luqmaan <luqmaan.badat@gmial.com> 
Date: Tue Jun 09 2020 14:59:49 GMT+0200 (SAST) 

    .gitignore


commit 98e06c8e 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Tue Jun 09 2020 14:38:54 GMT+0200 (SAST) 

    Skeleton Structure

We will be using poetry package dependency manager to handle our external libraries such as Vidgear and OpenCV.

This commit serves as the purpose of creating a structure for the home controller system such that other members of the team can work on a separate branch and use the same structure.

I have installed the foollowing packages:
* Black - formatting library to allign our python code to the industry standards.
* poetrify - used for compilling the dependencies into a requirements.txt that will be used to install our packages on deploy.

I have included a ReadMe that explains how to use poetry and integrate it with respective IDEs for development.