Updated S3 Integration to be generic no matter the conent to upload to S3
The process of uploading detected images and/or video content to S3 happens in the same fashion. All that you need to supply to the upload_to_s3 is the path to the path to the resource (locally), the file name, and the tag.

The current main is an example of uploading a detected image for the user demo1. The current photo is a detected owner :)

TODO: add confidential to store user_id persistently in the HCP in an encrypted manner