Login integration
Once the user logs into the HCP, two hidden files are created:
1. .hash that stores the users that have logged into the HCP on the same computer, and
2.  .current_user that stores the latest user that has logged into the HCP.

The .hash file is used as a dictionary of users in order to be able to keep track of the HCP ids of the user. Since the HCP_ID is created the first time the user logs into the HCP, that id needs to be persistently stored, otherwise, the HCP will not know which HCP belongs to the user on the machine that is running the HCP. This also accomodates multiuple users being able to log into one machine and not having conflicting HCP ids.

The HCP_id is a sha256 hash of the user's id and the timestamp that the user logged in. This is going to be persistently stored in the DB when the cameras are integrated with the API :+1: