Singleton User + Uploading of HCP & Cameras
I have created a singleton object User that is instantiated when the user logs in. If the user has previously logged in AND the .hash file exists in the root directory, it will retireve the HCP id of the user and store it in the singleton object.

Tokens are handled by the User such that everytime the user sends a request to the API, it checks if the token has expired, if so, it generates a new token. NOTE: by defualt AWSSRP generates a token that is valid for 1 Hour. Hence, after every hour, the token is generated again (TODO: use the refresh_token instead of generating a new one).

Now the HCP can integrate the cameras and login with The API.

TODO: add timestamp to the Video upload API route.