Video Upload & Error Handling
Exception handling in services if the user does not instantiate the singleton object BEFORE using the services. Hence, you NEED to provide the user_id, username, and password to authenticate the singleton user. Hence, this means that you NEED to use the login function by providing a valid username and password. I have provided tests that use a mock user with the following cerdentials:

Username := Foo
Password := Test@123

I have also integrated the services with the api

TODO: once the new api is deployed, I need to change the BASE_URL to reference the new stage in the API.