Squashed commit of the following
commit 25f122a1 
Author: JonathenSundy <JonathenSundy> 
Date: Fri Jul 10 2020 13:02:05 GMT+0200 (South Africa Standard Time) 

    Conflicts resolved in the console.

commit 0434d998 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Fri Jul 10 2020 12:55:22 GMT+0200 (South Africa Standard Time) 

    Video Upload & Error Handling

Exception handling in services if the user does not instantiate the singleton object BEFORE using the services. Hence, you NEED to provide the user_id, username, and password to authenticate the singleton user. Hence, this means that you NEED to use the login function by providing a valid username and password. I have provided tests that use a mock user with the following cerdentials:

Username := Foo
Password := Test@123

I have also integrated the services with the api

TODO: once the new api is deployed, I need to change the BASE_URL to reference the new stage in the API.


commit d21d2597 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jul 09 2020 22:17:20 GMT+0200 (South Africa Standard Time) 

    Singleton User + Uploading of HCP & Cameras

I have created a singleton object User that is instantiated when the user logs in. If the user has previously logged in AND the .hash file exists in the root directory, it will retireve the HCP id of the user and store it in the singleton object.

Tokens are handled by the User such that everytime the user sends a request to the API, it checks if the token has expired, if so, it generates a new token. NOTE: by defualt AWSSRP generates a token that is valid for 1 Hour. Hence, after every hour, the token is generated again (TODO: use the refresh_token instead of generating a new one).

Now the HCP can integrate the cameras and login with The API.

TODO: add timestamp to the Video upload API route.


commit 76a5f988 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jul 09 2020 00:29:26 GMT+0200 (South Africa Standard Time) 

    Login integration

Once the user logs into the HCP, two hidden files are created:
1. .hash that stores the users that have logged into the HCP on the same computer, and
2.  .current_user that stores the latest user that has logged into the HCP.

The .hash file is used as a dictionary of users in order to be able to keep track of the HCP ids of the user. Since the HCP_ID is created the first time the user logs into the HCP, that id needs to be persistently stored, otherwise, the HCP will not know which HCP belongs to the user on the machine that is running the HCP. This also accomodates multiuple users being able to log into one machine and not having conflicting HCP ids.

The HCP_id is a sha256 hash of the user's id and the timestamp that the user logged in. This is going to be persistently stored in the DB when the cameras are integrated with the API :+1:


commit 687cfe46 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jul 02 2020 21:00:18 GMT+0200 (South Africa Standard Time) 

    Updated S3 Integration to be generic no matter the conent to upload to S3

The process of uploading detected images and/or video content to S3 happens in the same fashion. All that you need to supply to the upload_to_s3 is the path to the path to the resource (locally), the file name, and the tag.

The current main is an example of uploading a detected image for the user demo1. The current photo is a detected owner :)

TODO: add confidential to store user_id persistently in the HCP in an encrypted manner


commit 49b2cc08 
Author: Jonathen Sundy <jonathensundy@gmail.com> 
Date: Thu Jul 02 2020 15:36:50 GMT+0200 (South Africa Standard Time) 

    Upload Detected Movement Images to S3

When the detected image from the HCP is identified, it will upload the image to S3. The S3 bucket has a trigger on this upload that will run the intruder detection module using rekognition.