Get Camera and Room/Location configuration
This achieves getting the camera configuration of a given HCP and the locations/rooms as a list of the names.

The upload_camera function has also been modified to include the path of the cameras in the db.

The configure() function is also called in the user class as it did not load the environment dictionary. Please revist this and fire the configure function wherever it is necessary.

The Prodcuer is also commented out such that it does not fire the listening of the socket stream when the user object is created. This functionality is specific and needs to be called in a function rather than on creation!

TODO:
Deletion functionality to allow for unit testing, and cleanup user class!