Updated camera upload convention
We do not need to incrementally add the site and location for a camera to be added, but rather need only add a camera with a given site id, location name, and camera id (with the respective metadata in the body as json).

Hence, the user need not upload the site when logging in for the first time as this is taken care of in the API.

TODO: change the way that the camera information is loaded up in the HCP. Since the location is not stored within the camera, and rather that each location has camras, it is the case that lodaing locations and cameras can be done in a one time fashion, instead of segmenting the data as we have been.