Updated the uploading to S3 for videos/images
The path to the API was incorrect and needed some adjustments to point to the correct endpoint.

In addition, the test_services file is updated sloghtly to allow for setting the generic user by using a helper fucntion, initialise_user. This did not change the Logic, but is made to make it easier for the user to understand what is happening and to remove rendundancy.